+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 11.0 [not yet released]
 
-- breaking change: fords and private roads are now always allowed and have to be excluded through the custom model, see e.g. custom_models/bike.json and #3076
+- replaced OSMHgvParser with OSMRoadAccessParser
+- breaking change: fords and certain roads (agricultural, forestry, private, delivery) are now always allowed for the `<vehicle>_access` encoded value and have to be excluded through the custom model, see e.g. custom_models/bike.json and #3076
 - use car_road_access instead of road_access. road_access exists only for web API backward compatibility. Note that MISSING was added in comparison to road_access, see #3066
 - road_access now contains value of highest transportation mode for CAR, i.e. access=private, motorcar=yes will now return YES and not PRIVATE
 - car.json by default avoids private roads